requirements.txt
flask
flask-limiter
python-dotenv
pymongo
bcrypt
pytz
dnspython
flask-sock
gunicorn
gevent
flask-wtf
stripe